requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
numpy==1.24.2
pedalboard==0.7.1
pycparser==2.21
sounddevice==0.4.6
typer==0.7.0
typing_extensions==4.4.0
pyyaml